% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{run_shiny}
\alias{run_shiny}
\title{make a quick shiny app to explore changing a function's parameters}
\usage{
run_shiny(call, slider_params, header = "", ...)
}
\arguments{
\item{call}{A character string, or a list of strings, each that evaluates to a call to a plotting function}

\item{slider_params}{A list of vectors providing the ranges for each parameter to make a slider for}

\item{header}{A character string for the app's title}
}
\value{
None.
}
\description{
make a quick shiny app to explore changing a function's parameters
}
\examples{
slider_params = list(A = c(1,0,0.5),B = c('a','b','c'))
plot_fun = function(A,B) {
              plot(NA,NA,xlim = c(-1,1),ylim = c(-1,1),main = B)
              symbols(0,0,circle = A,add=T,inches=F)
}
call = 'plot_fun(A,B)'
run_shiny(call,slider_params,header = 'test')
}

